requirements.txt
Django==2.0.5
django-redis==4.9.0
django-serializer==0.0.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
oauthlib==2.0.7
psycopg2==2.7.4
PyJWT==1.6.1
pytz==2018.4
redis==2.10.6
